project/users/requirements.txt
asyncpg==0.20.1
databases[postgresql]==0.2.6
fastapi==0.78.0
SQLAlchemy==1.3.13
uvicorn==0.11.2
httpx==0.11.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4